Catalogue # of galaxies from INT to LONG
The process of counting galaxies and creating a catalogue now uses
long variables instead of int since the number of galaxies might surpass
2.1 billion, which is near INT_MAX. We tested that the current version
returns the same catalogue as the previous one for ~1 million galaxies.
We also introduced a warning if NSIDE is greater than ~13000, which would
lead to Npixels>INT_MAX (Npixels remained int).

+++ src/corrlnfields_aux.cpp
@@ -48,7 +48,7 @@ int GetSubstrPos(const std::string field, const std::string header) {
 
 // Unless column=-1, write value to catalog[column][row] and update catSet:
 // Note that the catalog is transposed to ease FITS outputting.
-void CatalogFill(double **catalog, int row, int column, double value, int **catSet) {  
+void CatalogFill(double **catalog, long row, int column, double value, int **catSet) {  
   if(column==-1) return;
   
   // Write to catalog:            v Count number of updates in cell (for bookkeeping).
