+++ src/fitsfunctions.hpp
@@ -3,6 +3,6 @@
 
 #include "ParameterList.hpp"
 
-int WriteCatalog2Fits(std::string filename, double **table, int Nentries, const ParameterList & config);
+int WriteCatalog2Fits(std::string filename, double **table, long Nentries, const ParameterList & config);
 
 #endif
